Ensure that BitArrays are inferrable for arbitrary `Integer`
BitArray operations are frequently performed with partial inference
information (particularly from Pkg but others as well).
Currently the internal operations are a source of invalidation for
common arithmetic and bitwise operators. There seems to be no need
to live with this, as all the inference problems are fixed by
converting `Integer`s to `Int`s. This is done elsewhere already in
BitArray code, so it only seems to be continuing a process that is
already underway.

Because of the multiplicity of methods `BitArray{N}(a::Any)` is not
fully inferrable but there are at least limited victories,
and the real benefits are in the elimination of instabilities
in the internal operations.